Backport property_event using Dict storage
This one was tricky, because Godot 4.2 does not let you @export a Variant without specifying a variant type.
I was considering using get_property_list() with TYPE_NIL, but there's no documentation for this and it seemed like it had a high chance of not working.
However, Dictionaries can be serialized, and can store arbitrary Variants, so I made `value` and `initial_value` get/set a backing store Dictionary that stores a single key.
The catch is that this will absolutely break serialization when a project using the backport upgrades to Godot 4.3, so this needs to be made very clear for any upgraders:
(!!!) Upgrading from backport-4.2 to Godot 4.3 and godot-intervals/main WILL require manually updating the relevant .tres storage files. I'm OK with this personally because I know how to do that, but, it's a big gotcha. It's possible that there's some custom serialization support that could be added to PropertyEditor if 4.2 support was desired as a fundamental feature; value_store and initial_value_store could simply stick around for a few more versions, allowing the class to detect a version change for legacy 4.2 -> 4.3 serialization.